feat: add ARM template for Azure App Service deployment
Introduce a new Azure Resource Manager (ARM) template in JSON format that defines a deployment for an Azure App Service, including a resource group and an App Service Plan. Key parameters are specified with default values, and the template outlines the resources to be created along with their properties and dependencies.